backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

qdrant-client>=1.7.0

open-clip-torch>=2.24.0
torch>=2.0.0
pillow>=10.0.0
numpy>=1.24.0

deepface>=0.0.83
opencv-python-headless>=4.8.0

ultralytics>=8.0.0

google-generativeai>=0.8.0

pytesseract>=0.3.10

imagehash>=4.3.0
geopy>=2.4.0

pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
